requirements.txt

pytest==6.0.2
pytest-order==0.9.5
pytorch_lightning==1.2.3
pandas
numpy
optuna==2.3.0
fairscale==0.3.1
deepspeed==0.3.12
